Replace DELEGATE with DELEGATEE
Following @trishankatdatadog's and @jhdalek55's suggestion,
according to which, a delegate, when used as a noun, generally
implies a person. "Delegatee" is defined as "someone or something
to which something is delegated." In this case, when the
responsibility for signing is being granted to a different entity,
"delegatee" more clearly conveys this concept.